pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yakyoke"
version = "0.1.0"
description = "A local-first agent daemon. Bring your own LLM. Yokes the yak."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Robert Kempner" }]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "litellm>=1.40",
    "ddgs>=6.0",
    "trafilatura>=1.10",
    "python-dotenv>=1.0",
]

[project.scripts]
yakyoke = "yakyoke.cli:app"
yk = "yakyoke.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["yakyoke*"]
